エラーありのファイル作成とモジュールと関数をエクスポートしているファイルを作成
