+++ 03.asynchronous/promise.js
@@ -1,45 +1,4 @@
-import sqlite3 from "sqlite3";
-
-function createDatabase() {
-    const db = new sqlite3.Database(':memory:');
-    return db;
-}
-
-function runAsync(db,sql, values) {
-    return new Promise((resolve, reject) => {
-        db.run(sql, values, function (err) {
-            if (err) {
-                reject(err.message);
-            } else {
-                resolve(this.lastID);
-            }
-        })
-    })
-}
-
-function allAsync(db, sql) {
-    return new Promise((resolve, reject) => {
-        db.all(sql, (err, rows) => {
-            if (err) {
-                reject(err.message);
-            } else {
-                resolve(rows);
-            }
-        })
-    })
-}
-
-function closeAsync(db, sql) {
-	return new Promise((resolve, reject) => {
-		db.close((err) => {
-			if (err) {
-				reject(err.message);
-			} else {
-				resolve();
-			}
-		})
-	})
-}
+import { createDatabase, runAsync, allAsync, closeAsync } from "./async.js";
 
 const db = createDatabase();
 
@@ -65,5 +24,3 @@ runAsync(db, createTableQuery)
 	.then(() => {
 		return closeAsync(db)
 	})
-
-export {runAsync, allAsync, closeAsync};
